+++ _flake8_tergeo/hacks.py
@@ -1,13 +1,16 @@
+"""Hacks to support different python versions."""
+
 from __future__ import annotations
 
 import ast
 
 
-class DummyNode(ast.AST):
+class _DummyNode(ast.AST):
     pass
 
 
-def _add_python314_removed_ast_nodes():
-    ast.NameConstant = DummyNode
-    ast.Num = DummyNode
-    ast.Str = DummyNode
+def add_python314_removed_ast_nodes() -> None:
+    """Add dummy AST nodes for Python 3.14 removed nodes."""
+    ast.NameConstant = _DummyNode  # type:ignore[attr-defined]
+    ast.Num = _DummyNode  # type:ignore[attr-defined]
+    ast.Str = _DummyNode  # type:ignore[attr-defined]