+++ tests/checks/test_ast_imports.py
@@ -4,12 +4,10 @@ from __future__ import annotations
 
 import sys
 from functools import partial
-from typing import cast
 
 import pytest
 
 from _flake8_tergeo import Issue
-from _flake8_tergeo.checks import ast_import
 from tests.conftest import Flake8RunnerFixture
 
 FTP057 = partial(
@@ -46,6 +44,14 @@ FTP015 = partial(
     message=(
         "Found import of pkg_resources "
         "which should be replaced with a proper alternative of importlib.*"
+    ),
+)
+_FTP133 = partial(
+    Issue,
+    issue_number="FTP133",
+    message=(
+        "Using the compression namespace is recommended. "
+        "Replace the imported module with compression.{module}"
     ),
 )
 
@@ -85,6 +91,13 @@ def FTP030(  # pylint:disable=invalid-name
     return issue._replace(message=issue.message.format(future=future))
 
 
+def FTP133(  # pylint:disable=invalid-name
+    *, line: int, column: int, module: str
+) -> Issue:
+    issue = _FTP133(line=line, column=column)
+    return issue._replace(message=issue.message.format(module=module))
+
+
 def test_ftp001(runner: Flake8RunnerFixture) -> None:
     results = runner(filename="ftp001.txt", issue_number="FTP001")
     assert results == [
@@ -104,8 +117,17 @@ def test_ftp015(runner: Flake8RunnerFixture) -> None:
     assert results == [FTP015(line=8, column=1), FTP015(line=9, column=1)]
 
 
-def test_ftp030(runner: Flake8RunnerFixture) -> None:
-    results = runner(filename="ftp030.txt", issue_number="FTP030")
+@pytest.mark.parametrize(
+    "version,find_annotations", [("3.7.0", False), ("3.14.1", True)]
+)
+def test_ftp030(
+    runner: Flake8RunnerFixture, version: str, find_annotations: bool
+) -> None:
+    results = runner(
+        filename="ftp030.txt",
+        issue_number="FTP030",
+        args=("--ftp-python-version", version),
+    )
     assert results == [
         FTP030(line=2, column=1, future="nested_scopes"),
         FTP030(line=3, column=1, future="generators"),
@@ -115,6 +137,11 @@ def test_ftp030(runner: Flake8RunnerFixture) -> None:
         FTP030(line=7, column=1, future="print_function"),
         FTP030(line=8, column=1, future="unicode_literals"),
         FTP030(line=9, column=1, future="generator_stop"),
+        *(
+            [FTP030(line=12, column=1, future="annotations")]
+            if find_annotations
+            else []
+        ),
     ]
 
 
@@ -170,3 +197,39 @@ def test_ftp058(runner: Flake8RunnerFixture) -> None:
         FTP058(line=10, column=1),
         FTP058(line=11, column=1),
     ]
+
+
+@pytest.mark.parametrize(
+    "python_version,find_by_version", [("3.7.1", False), ("3.14.1", True)]
+)
+@pytest.mark.parametrize(
+    "imp,find_by_imp,module",
+    [
+        ("import bz2", True, "bz2"),
+        ("import gzip", True, "gzip"),
+        ("import lzma", True, "lzma"),
+        ("import zlib", True, "zlib"),
+        ("from zlib import some", True, "zlib"),
+        ("import zlib2", False, None),
+        ("from compression.gzip import some", False, None),
+        ("import compression.lzma", False, None),
+    ],
+)
+def test_ftp133(
+    runner: Flake8RunnerFixture,
+    python_version: str,
+    find_by_version: bool,
+    imp: str,
+    find_by_imp: bool,
+    module: str,
+) -> None:
+    results = runner(
+        filename="ftp133.txt",
+        issue_number="FTP133",
+        imp=imp,
+        args=("--ftp-python-version", python_version),
+    )
+    if find_by_version and find_by_imp:
+        assert results == [FTP133(line=1, column=1, module=module)]
+    else:
+        assert not results